almost compileable code generation for methods
