+++ tools/gen.R
@@ -182,14 +182,43 @@ case_dim_keepdim_and_other <- function(method) {
 
 argument_string <- function(argument) {
 
-  nullable <- argument$is_nullable
+  type <- argument$dynamic_type
 
-  glue::glue("{if (nullable) 'Rcpp::Nullable<' else ''}{argument$dynamic_type}{if (nullable) '>' else ''} {argument$name}")
+  if (type == "Tensor")
+    type <- "Rcpp::XPtr<torch::Tensor>"
+
+  if (type == "Scalar")
+    type <- "SEXP"
+
+  if (type == "IntArrayRef")
+    type <- "std::vector<std::int64_t>"
+
+  if (type == "ScalarType")
+    type <- "std::string"
+
+  if (type == "TensorList")
+    type <- "Rcpp::List"
+
+  if (argument$is_nullable)
+    type <- glue::glue("Rcpp::Nullable<{type}>")
+
+  if (type == "Generator *")
+    return(NA_character_) # remove generators from the call
+
+  if (type == "MemoryFormat")
+    return(NA_character_) # remove generators from the call
+
+  if (type == "ConstQuantizerPtr")
+    return(NA_character_) # remove generators from the call
+
+  glue::glue("{type} {argument$name}")
 }
 
 arguments_string <- function(arguments) {
-  arguments <- purrr::map_chr(arguments, argument_string)
-  paste(arguments, collapse = ", ")
+  arguments %>%
+    purrr::map_chr(argument_string) %>%
+    purrr::discard(is.na) %>%
+    paste(collapse = ", ")
 }
 
 return_type_string <- function(method) {
@@ -218,8 +247,36 @@ signature_string <- function(method) {
      ")
 }
 
+arguments_call_string <- function(argument) {
+
+  argument_name <- argument$name
+
+  if (argument$dynamic_type == "Scalar")
+    argument_name <- glue::glue("scalar_from_r_({argument_name})")
+
+  if (argument$dynamic_type == "ScalarType")
+    argument_name <- glue::glue("scalar_type_from_string({argument_name})")
+
+  if (argument$dynamic_type == "TensorList")
+    argument_name <- glue::glue("tensor_list_from_r_({argument_name})")
+
+  if (argument$dynamic_type == "Generator *")
+    return(NA_character_)
+
+  if (argument$dynamic_type == "MemoryFormat")
+    return(NA_character_)
+
+  if (argument$dynamic_type == "ConstQuantizerPtr")
+    return(NA_character_)
+
+  argument_name
+}
+
 arguments_preprocess <- function(arguments) {
-  paste(purrr::map_chr(arguments, ~.x$name), collapse = ", ")
+  arguments %>%
+    purrr::map_chr(arguments_call_string) %>%
+    purrr::discard(is.na) %>%
+    paste(collapse = ", ")
 }
 
 body_string <- function(method) {
@@ -283,15 +340,11 @@ method_string <- function(method) {
   }};")
 }
 
-
-for (name in method_names) {
-
-  method <- keep(tensor_methods, ~.x$name == name)
+make_method_code <- function(method) {
 
   if (length(method) == 1) {
 
-    print(method_string(method[[1]]))
-
+    method_string(method[[1]])
 
   } else {
 
@@ -310,6 +363,26 @@ for (name in method_names) {
   }
 
 }
+
+methods_code <- c()
+
+for (name in method_names) {
+  method <- keep(tensor_methods, ~.x$name == name)
+  methods_code <- c(methods_code, make_method_code(method))
+}
+
+if (file.exists(methods_fname))
+  file.remove(methods_fname)
+
+includes <- '// Dont modify!
+// This file is auto generated by tools/gen.R.
+#include "torch_types.h"
+#include "utils.hpp"
+#include "scalar.hpp"
+#include "device.hpp"
+'
+paste(c(includes, methods_code), collapse = "\n\n") %>%
+  writeLines(con = methods_fname)
